Initial design for server monitor app icon
Given that we're changing from a black app menu with white app icons to a white app menu with black app icons, I'm not sure what colour the icon should be. I've done it white, as that's what the icons have always been before, and I'm guessing we're going to do some code fu to modify them, but if you need it black, just let me know. It's a 2 second fix. ;)