Add NC section to serverinfo UI page
This adds a few monitoring information we already collect anyway - and were already available in the API endpoint.

This currently loads all SystemStatistics, which might be reduced to a smaller dataset in the future for performance reasons. Note: `webServer` in ApiController might be moved to `SystemStatistics` at a later point.

Signed-off-by: Patrik Kernstock <info@pkern.at>